backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "httpx>=0.27.0",
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "langchain-openai>=0.1.14",
    "openai>=1.6.1",
    "openpyxl==3.1.2",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "python-pptx==0.6.23",
    "slowapi==0.1.9",
    "uvicorn[standard]==0.24.0",
]
